Automate releases (#5)
This pull request defines a release workflow and Docker setup to
streamline the build and release process, including the addition of a
new GitHub Actions workflow for handling releases and updates to the
`Dockerfile` to support building and running the application in a Docker
container.

The objective of this release configuration is to maintain a 'latest'
release, as well as tagged releases at milestones.

It's worth noting that the binary as built at the moment is unlikely to
remain, as [Deno does not handle dynamic imports when
compiled](https://github.com/denoland/deno/issues/18327). I'm looking
into fallbacks for this, which are likely to include a bootstrapping
script which will install Deno and then run Trove locally. This will
apply to non-Docker installs only, since the Docker installs will simply
run Trove within the container.