Add Bootstrap script, support config loading from multiple sources (#14)
This pull request introduces a bootstrap script which can be used to get
Trove running quickly. The bootstrap script ensures that Deno is
installed, and downloads it to a specific directory if not. It then runs
Trove using a remote import with Deno, or alternatively supports running
either HEAD (main), or a specific branch via a git clone.

This makes it very easy to run Trove for people who don't want to use
the Docker image. It also provides a recipe for those that might not
want to install and/or run via a shell script.

Along with the bootstrap script, this PR also improves config and plugin
loading to work with relative, absolute, and HTTPS imports. Previously,
we had some problems with config importing outside the Trove repository
folder, due to how Deno treats `Deno.pwd` as being derived based on the
script path, rather than the shell PWD. This change means that now
config and plugins can be loaded using any version of relative or
absolute paths safely.